requirements.txt
annotated-types==0.5.0
anyio==3.7.1
borb==2.1.17
boto3==1.28.73
botocore==1.31.73
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
cryptography==41.0.4
et-xmlfile==1.1.0
fastapi==0.103.1
fonttools==4.43.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.1
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
jmespath==1.0.1
lxml==4.9.3
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyPDF2==3.0.1
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.8.2
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
